refactor: unify CLI agent backends behind one invoker, relocate runner
The two backends duplicated the entire invocation flow: build args,
type-assert StreamingRunner, stream with a per-line log decoder, fall
back to stdin piping, salvage cost on failure, wrap the error with
stderr. Both copies now delegate to one runCLIAgent invoker driven by a
per-backend spec {binary, args, log decoder, parser}, so the semantics
(streaming, ARG_MAX-safe stdin prompts, cost-only salvage) are defined
exactly once.

File organization untangled to match names:
- runner.go (new): CommandRunner/StreamingRunner/ExecRunner,
  BuildAgentEnv - this infrastructure lived in claude.go
- agent.go: CodeAgent interface + the shared invoker
- claude.go: ClaudeCodeAgent and its stream parsing (the type formerly
  lived in agent.go while claude.go held the runner)
- opencode.go: OpenCodeAgent and its parsing, unchanged location

Also deduplicated within the runner: Run now delegates to RunWithStdin,
and the env-overlay/command construction shared by all three run
methods is one helper with the last-wins env semantics documented.
The 200-byte log preview truncation existed in both backends as raw
byte slicing that could split UTF-8 sequences; it is now one rune-safe
previewText helper.
